Display class created. Input Handled nicely.
