+++ Display.h
@@ -1,4 +1,5 @@
-#pragma once
+#ifndef DISPLAY_H
+#define DISPLAY_H
 #include <glad/glad.h>
 #include <GLFW/glfw3.h>
 #include <iostream>
@@ -14,44 +15,6 @@
 	- Initialize GLAD
 
 */
-
-struct InputContainer {
-	int W;
-	int A;
-	int S;
-	int D;
-	int SPACE;
-	int ESCAPE;
-} gInput;
-
-void KeyCallback(GLFWwindow*, int key, int scancode, int action, int mods) {
-	switch (key) {
-	case GLFW_KEY_W:
-		gInput.W = action;
-		break;
-	case GLFW_KEY_A:
-		gInput.A = action;
-		break;
-	case GLFW_KEY_S:
-		gInput.S = action;
-		break;
-	case GLFW_KEY_D:
-		gInput.D = action;
-		break;
-	case GLFW_KEY_SPACE:
-		gInput.SPACE = action;
-		break;
-	case GLFW_KEY_ESCAPE:
-		gInput.ESCAPE = action;
-		break;
-	default:
-		break;
-	}
-}
-
-void ErrorCallback(int error, const char* description) {
-	std::cout << "Error: " << description << std::endl;
-}
 
 class Display {
 public:
@@ -69,3 +32,4 @@ private:
 	int mWidth, mHeight;
 	std::string mTitle;
 };
+#endif
